Add training data
For now I will use just one dataset for testing my algorithms: the
popular wine quality data. In this commit, I retrieve this data and
the associated descriptions of the data.

There are two datasets, one for red wine and one for white. Each has 11
input variables and one output variable (quality). The goal is, of
course, to predict quality.